style(frontend): move rtk-queries to redux/api
