+++ frontend/src/screens/DriverPlanRide/DriverPlanRide.tsx
@@ -10,9 +10,9 @@ import RouteMarker from '~/components/maps/RouteMarker'
 import MapViewWithRoute from '~/components/MapViewWithRoute'
 import TopNavBar from '~/components/nav/TopNavBar'
 import { useCurrentLocation } from '~/hooks/useCurrentLocation'
-import { useCreateRideMutation } from '~/redux/driver'
-import { useGetRouteQuery } from '~/redux/maps'
-import { useGetSavedRidesQuery } from '~/redux/users'
+import { useCreateRideMutation } from '~/redux/api/driver'
+import { useGetRouteQuery } from '~/redux/api/maps'
+import { useGetSavedRidesQuery } from '~/redux/api/users'
 import { type DriverPlanRideScreenProps } from '~/types/screens'
 import { isValidWaypoint, isValidWaypoints } from '~/utils/maps'
