+++ frontend/src/screens/DriverSelectJoin/DriverSelectJoin.tsx
@@ -7,8 +7,8 @@ import { skipToken } from '@reduxjs/toolkit/query'
 import MapViewWithRoute from '~/components/MapViewWithRoute'
 import { ConfirmModal } from '~/components/modals/Confirm'
 import TopNavBar from '~/components/nav/TopNavBar'
-import { useGetJoinsQuery, useGetRideQuery, useRespondJoinMutation } from '~/redux/driver'
-import { useGetCurrentActivityQuery } from '~/redux/users'
+import { useGetJoinsQuery, useGetRideQuery, useRespondJoinMutation } from '~/redux/api/driver'
+import { useGetCurrentActivityQuery } from '~/redux/api/users'
 import { type DriverSelectJoinScreenProps } from '~/types/screens'
 
 import { PassengerAvatarList, PassengerCardList } from './PassengerList'
